refactor: rename utils.js file and move SOURCE_FOLDER constant to it
