+++ src/fs/constants.js
@@ -1,5 +1,9 @@
 import { dirname } from 'path';
 import { fileURLToPath } from 'url';
 
+const SOURCE_FOLDER = 'files';
+
 const fileName = fileURLToPath(import.meta.url);
-export const __dirName = dirname(fileName);
+const __dirName = dirname(fileName);
+
+export { __dirName, SOURCE_FOLDER };
