+++ src/fs/rename.js
@@ -1,12 +1,12 @@
 import * as fs from 'fs/promises';
-import { __dirName } from './utils.js';
+import { __dirName, SOURCE_FOLDER } from './constants.js';
 
-const SOURCE_FILE = 'files/wrongFilename.txt';
-const RENAMED_FILE = 'files/properFilename.md';
+const SOURCE_FILE = 'wrongFilename.txt';
+const RENAMED_FILE = 'properFilename.md';
 const ERROR_MSG = 'FS operation failed';
 
-const source_path = __dirName + '/' + SOURCE_FILE;
-const renamed_path = __dirName + '/' + RENAMED_FILE;
+const source_path = `${__dirName}/${SOURCE_FOLDER}/${SOURCE_FILE}`;
+const renamed_path = `${__dirName}/${SOURCE_FOLDER}/${RENAMED_FILE}`;
 
 const rename = async () => {
     try {
